Update Char conversion to be toInt().toChar()
Due to the deprecation of Long.toChar() as part of KEEP-0227: char-int-conversions

Bug: 479862364
Test: ./gradlew :room3:room3-compiler:test

+++ room3/room3-compiler/src/test/test-data/kotlinCodeGen/dataClassRowAdapter_primitives_nullable.kt
@@ -125,7 +125,7 @@ internal class MyDao_Impl(
           if (_stmt.isNull(_columnIndexOfChar)) {
             _tmpChar = null
           } else {
-            _tmpChar = _stmt.getLong(_columnIndexOfChar).toChar()
+            _tmpChar = _stmt.getLong(_columnIndexOfChar).toInt().toChar()
           }
           val _tmpFloat: Float?
           if (_stmt.isNull(_columnIndexOfFloat)) {
